[refactoring] naming과 package 변경
