+++ src/main/java/racing/view/Input.java
@@ -1,6 +1,6 @@
 package racing.view;
 
-import racing.domain.InputBuilder;
+import racing.domain.GetNameList;
 
 import java.util.List;
 import java.util.Scanner;
@@ -15,8 +15,8 @@ public class Input {
     public static List<String> getNameList() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         String nameString = scanner.nextLine();
-        InputBuilder inputBuilder = new InputBuilder(nameString);
-        List<String> nameList = inputBuilder.getNameList();
+        GetNameList getNameList = new GetNameList(nameString);
+        List<String> nameList = getNameList.getNameList();
         return nameList;
     }
